Broke up functions, changed back to JSON stringify for dict deep copy
